+++ pysatNASA/instruments/methods/__init__.py
@@ -4,6 +4,11 @@ from pysatNASA.instruments.methods._cdf import CDF  # noqa F401
 from pysatNASA.instruments.methods import cdaweb  # noqa F401
 from pysatNASA.instruments.methods import cnofs  # noqa F401
 from pysatNASA.instruments.methods import de2  # noqa F401
+from pysatNASA.instruments.methods import dmsp  # noqa F401
 from pysatNASA.instruments.methods import general  # noqa F401
+from pysatNASA.instruments.methods import gps  # noqa F401
 from pysatNASA.instruments.methods import icon  # noqa F401
+from pysatNASA.instruments.methods import jhuapl  # noqa F401
 from pysatNASA.instruments.methods import omni  # noqa F401
+from pysatNASA.instruments.methods import ses14  # noqa F401
+from pysatNASA.instruments.methods import timed  # noqa F401