+++ pysatNASA/instruments/methods/cdaweb.py
@@ -9,6 +9,7 @@ Adding new CDAWeb datasets should only require mininal user intervention.
 
 import cdflib
 import datetime as dt
+import numpy as np
 import os
 import pandas as pds
 import requests
@@ -16,6 +17,7 @@ import warnings
 import xarray as xr
 
 from bs4 import BeautifulSoup
+from cdasws import CdasWs
 
 import pysat
 from pysat.instruments.methods import general
@@ -31,6 +33,35 @@ except ImportError:
     auto_CDF = libCDF
 
 
+def try_inst_dict(inst_id, tag, supported_tags):
+    """Check that the inst_id and tag combination is valid.
+
+    Parameters
+    ----------
+    tag : str
+        Data product tag (default='')
+    inst_id : str
+        Instrument ID (default='')
+    supported_tags : dict
+        dict of dicts. Keys are supported tag names for download. Value is
+        a dict with 'remote_dir', 'fname'. Inteded to be
+        pre-set with functools.partial then assigned to new instrument code.
+        (default=None)
+
+    Returns
+    -------
+    inst_dict : dict or str
+        dictionary containing file location in spdf archive, or dataset ID for
+        cdasws
+    """
+    try:
+        inst_dict = supported_tags[inst_id][tag]
+    except KeyError:
+        raise ValueError('inst_id / tag combo unknown.')
+
+    return inst_dict
+
+
 def load(fnames, tag='', inst_id='', file_cadence=dt.timedelta(days=1),
          flatten_twod=True, pandas_format=True, epoch_name='Epoch',
          meta_processor=None, meta_translation=None, drop_meta_labels=None,
@@ -42,9 +73,9 @@ def load(fnames, tag='', inst_id='', file_cadence=dt.timedelta(days=1),
     fnames : pandas.Series
         Series of filenames
     tag : str
-        tag or None (default='')
+        Data product tag (default='')
     inst_id : str
-        satellite id or None (default='')
+        Instrument ID (default='')
     file_cadence : dt.timedelta or pds.DateOffset
         pysat assumes a daily file cadence, but some instrument data files
         contain longer periods of time.  This parameter allows the specification
@@ -93,8 +124,10 @@ def load(fnames, tag='', inst_id='', file_cadence=dt.timedelta(days=1),
 
     Note
     ----
-    This routine is intended to be used by pysat instrument modules supporting
-    a particular NASA CDAWeb dataset
+    - This routine is intended to be used by pysat instrument modules supporting
+      a particular NASA CDAWeb dataset
+    - pysatCDF (as of v0.3.2) does not support numpy>=1.24.  Load errors may
+      arise.  See https://github.com/pysat/pysatCDF/issues/46
 
     """
 
@@ -110,6 +143,7 @@ def load(fnames, tag='', inst_id='', file_cadence=dt.timedelta(days=1),
 
         data, meta = load_xarray(fnames, tag=tag, inst_id=inst_id,
                                  epoch_name=epoch_name,
+                                 file_cadence=file_cadence,
                                  meta_processor=meta_processor,
                                  meta_translation=meta_translation,
                                  drop_meta_labels=drop_meta_labels)
@@ -125,9 +159,9 @@ def load_pandas(fnames, tag='', inst_id='', file_cadence=dt.timedelta(days=1),
     fnames : pandas.Series
         Series of filenames
     tag : str
-        tag or None (default='')
+        Data product tag (default='')
     inst_id : str
-        satellite id or None (default='')
+        Instrument ID (default='')
     file_cadence : dt.timedelta or pds.DateOffset
         pysat assumes a daily file cadence, but some instrument data files
         contain longer periods of time.  This parameter allows the specification
@@ -224,13 +258,12 @@ def load_pandas(fnames, tag='', inst_id='', file_cadence=dt.timedelta(days=1),
 
 
 def load_xarray(fnames, tag='', inst_id='',
-                labels={'units': ('units', str), 'name': ('long_name', str),
-                        'notes': ('notes', str), 'desc': ('desc', str),
-                        'plot': ('plot_label', str), 'axis': ('axis', str),
-                        'scale': ('scale', str),
-                        'min_val': ('value_min', float),
-                        'max_val': ('value_max', float),
-                        'fill_val': ('fill', float)},
+                file_cadence=dt.timedelta(days=1),
+                labels={'units': ('Units', str), 'name': ('Long_Name', str),
+                        'notes': ('Var_Notes', str), 'desc': ('CatDesc', str),
+                        'min_val': ('ValidMin', float),
+                        'max_val': ('ValidMax', float),
+                        'fill_val': ('FillVal', float)},
                 epoch_name='Epoch', meta_processor=None,
                 meta_translation=None, drop_meta_labels=None):
     """Load NASA CDAWeb CDF files into an xarray Dataset.
@@ -240,9 +273,14 @@ def load_xarray(fnames, tag='', inst_id='',
     fnames : pandas.Series
         Series of filenames
     tag : str
-        tag or (default='')
+        Data product tag (default='')
     inst_id : str
-        satellite id (default='')
+        Instrument ID (default='')
+    file_cadence : dt.timedelta or pds.DateOffset
+        pysat assumes a daily file cadence, but some instrument data files
+        contain longer periods of time.  This parameter allows the specification
+        of regular file cadences greater than or equal to a day (e.g., weekly,
+        monthly, or yearly). (default=dt.timedelta(days=1))
     labels : dict
         Dict where keys are the label attribute names and the values are tuples
         that have the label values and value types in that order.
@@ -308,7 +346,16 @@ def load_xarray(fnames, tag='', inst_id='',
         # metadata for pysat using some assumptions. Depending upon your needs
         # the resulting pandas DataFrame may need modification.
         ldata = []
-        for lfname in fnames:
+
+        # Find unique files for monthly / yearly cadence.
+        # Arbitrary timestamp needed for comparison.
+        t0 = dt.datetime(2009, 1, 1)
+        if (t0 + file_cadence) > (t0 + dt.timedelta(days=1)):
+            lfnames = list(np.unique([fname[:-11] for fname in fnames]))
+        else:
+            lfnames = fnames
+
+        for lfname in lfnames:
             temp_data = cdflib.cdf_to_xarray(lfname, to_datetime=True)
             ldata.append(temp_data)
 
@@ -396,9 +443,10 @@ def load_xarray(fnames, tag='', inst_id='',
     return data, meta
 
 
+# TODO(#103): Include support to unzip / untar files after download.
 def download(date_array, tag='', inst_id='', supported_tags=None,
              remote_url='https://cdaweb.gsfc.nasa.gov', data_path=None):
-    """Download NASA CDAWeb CDF data.
+    """Download NASA CDAWeb data.
 
     This routine is intended to be used by pysat instrument modules supporting
     a particular NASA CDAWeb dataset.
@@ -408,9 +456,9 @@ def download(date_array, tag='', inst_id='', supported_tags=None,
     date_array : array-like
         Array of datetimes to download data for. Provided by pysat.
     tag : str
-        tag or None (default='')
+        Data product tag (default='')
     inst_id : str
-        satellite id or None (default='')
+        Instrument ID (default='')
     supported_tags : dict
         dict of dicts. Keys are supported tag names for download. Value is
         a dict with 'remote_dir', 'fname'. Inteded to be pre-set with
@@ -427,7 +475,7 @@ def download(date_array, tag='', inst_id='', supported_tags=None,
     --------
     ::
 
-        # download support added to cnofs_vefi.py using code below
+        # Download support added to cnofs_vefi.py using code below
         fn = 'cnofs_vefi_bfield_1sec_{year:4d}{month:02d}{day:02d}_v05.cdf'
         dc_b_tag = {'remote_dir': ''.join(('/pub/data/cnofs/vefi/bfield_1sec',
                                             '/{year:4d}/')),
@@ -439,10 +487,7 @@ def download(date_array, tag='', inst_id='', supported_tags=None,
 
     """
 
-    try:
-        inst_dict = supported_tags[inst_id][tag]
-    except KeyError:
-        raise ValueError('inst_id / tag combo unknown.')
+    inst_dict = try_inst_dict(inst_id, tag, supported_tags)
 
     # Naming scheme for files on the CDAWeb server
     remote_dir = inst_dict['remote_dir']
@@ -455,14 +500,23 @@ def download(date_array, tag='', inst_id='', supported_tags=None,
                                      stop=date_array[-1])
 
     # Download only requested files that exist remotely
-    for date, fname in remote_files.iteritems():
+    for date, fname in remote_files.items():
         # Format files for specific dates and download location
-        formatted_remote_dir = remote_dir.format(year=date.year,
-                                                 month=date.month,
-                                                 day=date.day,
-                                                 hour=date.hour,
-                                                 min=date.minute,
-                                                 sec=date.second)
+        # Year and day found in remote_dir: day is assumed to be day of year
+        if 'day' in remote_dir and 'month' not in remote_dir:
+            doy = date.timetuple().tm_yday
+            formatted_remote_dir = remote_dir.format(year=date.year,
+                                                     day=doy,
+                                                     hour=date.hour,
+                                                     min=date.minute,
+                                                     sec=date.second)
+        else:
+            formatted_remote_dir = remote_dir.format(year=date.year,
+                                                     month=date.month,
+                                                     day=date.day,
+                                                     hour=date.hour,
+                                                     min=date.minute,
+                                                     sec=date.second)
         remote_path = '/'.join((remote_url.strip('/'),
                                 formatted_remote_dir.strip('/'),
                                 fname))
@@ -488,7 +542,102 @@ def download(date_array, tag='', inst_id='', supported_tags=None,
     return
 
 
-def list_remote_files(tag=None, inst_id=None, start=None, stop=None,
+def cdas_download(date_array, tag='', inst_id='', supported_tags=None,
+                  data_path=None):
+    """Download NASA CDAWeb CDF data using cdasws.
+
+    This routine is intended to be used by pysat instrument modules supporting
+    a particular NASA CDAWeb dataset.
+
+    Parameters
+    ----------
+    date_array : array-like
+        Array of datetimes to download data for. Provided by pysat.
+    tag : str
+        Data product tag (default='')
+    inst_id : str
+        Instrument ID (default='')
+    supported_tags : dict
+        dict of dicts. Keys are supported tag names for download. Value is
+        a dict with 'remote_dir', 'fname'. Inteded to be pre-set with
+        functools.partial then assigned to new instrument code.
+        (default=None)
+    data_path : str or NoneType
+        Path to data directory.  If None is specified, the value previously
+        set in Instrument.files.data_path is used.  (default=None)
+
+    Note
+    ----
+    Supported tags for this function use the cdaweb dataset naming convention.
+    You can find the data set names on CDAWeb or you can use cdasws.
+
+    Starting from scratch using cdasws
+    ::
+        from cdasws import CdasWs
+        cdas = CdasWs()
+
+        # Get list of available observatories/platforms.
+        cdas.get_observatories()
+
+        # Once your observatory is located, list the available instruments.
+        cdas.get_instruments(observatory=‘observatory_name’)
+
+        # Now list the available data sets for one instrument.
+        cdas.get_datasets(observatory=‘observatory_name’,
+                          instrument=‘instrument_name’)
+
+        # You can also list all of the datasets for an observatory.
+        cdas.get_datasets(observatory=‘observatory_name’)
+
+    Alternatively
+    ::
+        Visit https://cdaweb.gsfc.nasa.gov/
+        Select the observatory you want from the list and press submit.
+        The following page will have a list of the data sets.
+        The bolded names are in the format that cdasws uses.
+
+    Examples
+    --------
+    ::
+        # Download support added to cnofs_vefi.py using code below
+        download_tags = {'': {'dc_b': 'CNOFS_VEFI_BFIELD_1SEC'}}
+        download = functools.partial(cdw.cdas_download,
+                                     supported_tags=download_tags)
+
+    """
+
+    start = date_array[0]
+    stop = date_array[-1]
+    remote_files = cdas_list_remote_files(tag=tag, inst_id=inst_id,
+                                          start=start, stop=stop,
+                                          supported_tags=supported_tags,
+                                          series_out=False)
+
+    for file in remote_files:
+
+        fname = file.split('/')[-1]
+        saved_local_fname = os.path.join(data_path, fname)
+
+        # Perform download
+        logger.info(' '.join(('Attempting to download file: ',
+                              file)))
+        try:
+            with requests.get(file) as req:
+                if req.status_code != 404:
+                    with open(saved_local_fname, 'wb') as open_f:
+                        open_f.write(req.content)
+                    logger.info('Successfully downloaded {:}.'.format(
+                        saved_local_fname))
+                else:
+                    logger.info(' '.join(('File: "', file,
+                                          '" is not available')))
+        except requests.exceptions.RequestException as exception:
+            logger.info(' '.join((str(exception), '- File: "', file,
+                                  '" Is not available')))
+    return
+
+
+def list_remote_files(tag='', inst_id='', start=None, stop=None,
                       remote_url='https://cdaweb.gsfc.nasa.gov',
                       supported_tags=None, two_digit_year_break=None,
                       delimiter=None):
@@ -500,11 +649,9 @@ def list_remote_files(tag=None, inst_id=None, start=None, stop=None,
     Parameters
     ----------
     tag : str
-        Denotes type of file to load.  Accepted types are <tag strings>.
-        (default='')
+        Data product tag (default='')
     inst_id : str
-        Specifies the satellite ID for a constellation.
-        (default='')
+        Instrument ID (default='')
     start : dt.datetime or NoneType
         Starting time for file list. A None value will start with the first
         file found.
@@ -553,10 +700,7 @@ def list_remote_files(tag=None, inst_id=None, start=None, stop=None,
 
     """
 
-    try:
-        inst_dict = supported_tags[inst_id][tag]
-    except KeyError:
-        raise ValueError('inst_id / tag combo unknown.')
+    inst_dict = try_inst_dict(inst_id, tag, supported_tags)
 
     # Naming scheme for files on the CDAWeb server
     format_str = '/'.join((inst_dict['remote_dir'].strip('/'),
@@ -601,18 +745,27 @@ def list_remote_files(tag=None, inst_id=None, start=None, stop=None,
         stop = dt.datetime.now() if (stop is None) else stop
 
         if 'year' in search_dir['keys']:
+            url_list = []
             if 'month' in search_dir['keys']:
                 search_times = pds.date_range(start,
                                               stop + pds.DateOffset(months=1),
                                               freq='M')
+                for time in search_times:
+                    subdir = format_dir.format(year=time.year, month=time.month)
+                    url_list.append('/'.join((remote_url, subdir)))
             else:
-                search_times = pds.date_range(start,
-                                              stop + pds.DateOffset(years=1),
-                                              freq='Y')
-            url_list = []
-            for time in search_times:
-                subdir = format_dir.format(year=time.year, month=time.month)
-                url_list.append('/'.join((remote_url, subdir)))
+                if 'day' in search_dir['keys']:
+                    search_times = pds.date_range(start, stop
+                                                  + pds.DateOffset(days=1),
+                                                  freq='D')
+                else:
+                    search_times = pds.date_range(start, stop
+                                                  + pds.DateOffset(years=1),
+                                                  freq='Y')
+                for time in search_times:
+                    doy = int(time.strftime('%j'))
+                    subdir = format_dir.format(year=time.year, day=doy)
+                    url_list.append('/'.join((remote_url, subdir)))
     try:
         for top_url in url_list:
             for level in range(n_layers + 1):
@@ -656,3 +809,82 @@ def list_remote_files(tag=None, inst_id=None, start=None, stop=None,
         stored_list = stored_list[mask]
 
     return stored_list
+
+
+def cdas_list_remote_files(tag='', inst_id='', start=None, stop=None,
+                           supported_tags=None, series_out=True):
+    """Return a list of every file for chosen remote data.
+
+    This routine is intended to be used by pysat instrument modules supporting
+    a particular NASA CDAWeb dataset.
+
+    Parameters
+    ----------
+    tag : str
+        Data product tag (default='')
+    inst_id : str
+        Instrument ID (default='')
+    start : dt.datetime or NoneType
+        Starting time for file list. A None value will start with the first
+        file found.
+        (default=None)
+    stop : dt.datetime or NoneType
+        Ending time for the file list.  A None value will stop with the last
+        file found.
+        (default=None)
+    supported_tags : dict
+        dict of dicts. Keys are supported tag names for download. Value is
+        a dict with 'remote_dir', 'fname'. Inteded to be
+        pre-set with functools.partial then assigned to new instrument code.
+        (default=None)
+    series_out : bool
+        boolean to determine output type. True for pandas series of file names,
+        and False for a list of the full web address.
+
+    Returns
+    -------
+    file_list : list
+        A list containing the verified available files
+
+    Note
+    ----
+    Supported tags for this function use the cdaweb dataset naming convention.
+    You can find the dataset names on cdaweb or you can use cdasws.
+
+    Examples
+    --------
+    ::
+        download_tags = {'': {'dc_b': 'CNOFS_VEFI_BFIELD_1SEC'}}
+        list_remote_files = functools.partial(cdw.cdas_list_remote_files,
+                                              supported_tags=download_tags)
+
+        download_tags = {'': {'': 'CNOFS_CINDI_IVM_500MS'}}
+        list_remote_files = functools.partial(cdw.cdas_list_remote_files,
+                                              supported_tags=download_tags)
+    """
+    cdas = CdasWs()
+    dataset = try_inst_dict(inst_id, tag, supported_tags)
+
+    if start is None and stop is None:
+        # Use the topmost directory without variables
+        start = cdas.get_inventory(identifier=dataset)[0].start
+        stop = cdas.get_inventory(identifier=dataset)[-1].end
+    elif stop is None:
+        stop = start + dt.timedelta(days=1)
+    elif start == stop:
+        stop = start + dt.timedelta(days=1)
+
+    if isinstance(start, pds._libs.tslibs.timestamps.Timestamp):
+        start = start.tz_localize('utc')
+        stop = stop.tz_localize('utc')
+
+    og_files = cdas.get_original_files(dataset=dataset, start=start, end=stop)
+
+    if series_out:
+        name_list = [os.path.basename(f['Name']) for f in og_files[1]]
+        t_stamp = [pds.Timestamp(f['StartTime'][:10]) for f in og_files[1]]
+        file_list = pds.Series(data=name_list, index=t_stamp)
+    else:
+        file_list = [f['Name'] for f in og_files[1]]
+
+    return file_list