+++ CHANGELOG.md
@@ -2,7 +2,46 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
-## [0.0.4] - 2022-XX-XX
+## [0.X.X] - 2023-XX-XX
+* New Instruments
+  * ACE EPAM
+  * ACE MAG
+  * ACE SIS
+  * ACE SWEPAM
+  * DE2 Fabry-Perot Interferometer (FPI)
+  * DE2 Vector Electric Field Instrument (VEFI) and magnetometer
+  * DMSP SSUSI EDR-Aurora data
+  * TIMED GUVI
+* Add TIMED GUVI platform to support L1C intensity datasets.
+  * Type of sensor source handled by inst_id with options of
+    spectrograph, imaging
+  * Resolution of dataset handled by tag with
+    low, high
+* Added CDAWeb methods that can use cdasws to get the remote file list
+* Bug Fixes
+  * Updated CDAWeb routines to allow for data stored by year/day-of-year
+  * Updated GOLD nmax to sort scans by time.
+  * Added 1 usec to GOLD nmax channel B times to ensure uniqueness
+* Documentation
+  * Added TIMED-GUVI platform
+  * Added missing sub-module imports
+  * Added discussion of ICON constellation to docstrings, including caveats
+* Enhancements
+  * Updated platform methods to follow a consistent style and work with the
+    general `init` function
+  * Added unit tests for the different platform method attributes
+  * xarray support for TIMED SEE
+* Maintenance
+  * Removed duplicate tests if pysatCDF not isntalled
+  * Only test pysatCDF on GitHub Actions for older numpy versions
+  * Updated actions and templates based on pysatEcosystem docs
+  * Remove pandas cap on NEP29 tests
+  * Updated dosctring style for consistency
+  * Removed version cap for xarray
+  * Added manual workflow to check that latest RC is installable through test pip
+  * Use pyproject.toml to manage setup
+
+## [0.0.4] - 2022-11-07
 * Update instrument tests with new test class
 * Support xarray datasets through cdflib
 * Preferentially loads data into pandas using pysatCDF if installed
@@ -22,7 +61,10 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   * Move OMNI HRO custom functions to a methods module
   * Deprecate OMNI HRO custom functions in instrument module
   * Update GitHub actions to the latest versions
-  * Use pyproject.toml to manage setup
+  * Added downstream test to test code with pysat RC
+  * Remove deprecated `convert_timestamp_to_datetime` calls
+  * Remove deprecated pandas syntax
+  * Added version cap for xarray 2022.11
 * Documentation
   * New logo added